Move ExceptionUtil out of util to an infrastructure module.
