+++ infrastructure/exceptions/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+  implementation project(':infrastructure:async')
+
+  implementation 'org.apache.commons:commons-lang3'
+}
