+++ infrastructure/exceptions/src/main/java/tech/pegasys/teku/infrastructure/exceptions/ExceptionUtil.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.exceptions;
+package tech.pegasys.teku.infrastructure.exceptions;
 
 import java.util.Optional;
 import org.apache.commons.lang3.exception.ExceptionUtils;
