+++ infrastructure/exceptions/src/test/java/tech/pegasys/teku/infrastructure/exceptions/ExceptionUtilTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.exceptions;
+package tech.pegasys.teku.infrastructure.exceptions;
 
 import static org.assertj.core.api.Assertions.assertThat;
