+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/gossip/topics/topichandlers/Eth2TopicHandler.java
@@ -29,7 +29,7 @@ import tech.pegasys.teku.networking.p2p.gossip.PreparedGossipMessage;
 import tech.pegasys.teku.networking.p2p.gossip.TopicHandler;
 import tech.pegasys.teku.ssz.SSZTypes.Bytes4;
 import tech.pegasys.teku.statetransition.validation.InternalValidationResult;
-import tech.pegasys.teku.util.exceptions.ExceptionUtil;
+import tech.pegasys.teku.infrastructure.exceptions.ExceptionUtil;
 
 public class Eth2TopicHandler<T> implements TopicHandler {
   private static final Logger LOG = LogManager.getLogger();
