+++ settings.gradle
@@ -21,6 +21,7 @@ include 'infrastructure:async'
 include 'infrastructure:collections'
 include 'infrastructure:crypto'
 include 'infrastructure:events'
+include 'infrastructure:exceptions'
 include 'infrastructure:http'
 include 'infrastructure:io'
 include 'infrastructure:metrics'
