+++ teku/build.gradle
@@ -3,6 +3,7 @@ dependencies {
   implementation project(':data')
   implementation project(':data:serializer')
   implementation project(':data:beaconrestapi')
+  implementation project(':infrastructure:exceptions')
   implementation project(':infrastructure:metrics')
   implementation project(':data:dataexchange')
   implementation project(':data:yaml')
